try to manage different times format between fopi and pof

+++ app/services/time_utils.py
@@ -20,9 +20,12 @@ def calculate_time_index(ds, index: str, base_time: str, lead_hours: int) -> int
     """
     base_file_time = extract_base_time_from_encoding(ds, index)
     valid_hours = compute_valid_hour(base_file_time, base_time, lead_hours)
+
     logger.info(f"ℹ️ index in calculate_time_index: {index}")
+
     if index == "fopi":
-        time_in_hours = ds.time.values  # already float
+        # Explicitly cast to float to avoid dtype issues
+        time_in_hours = ds.time.values.astype(float)
     else:
         time_in_hours = np.array([
             (pd.to_datetime(t).to_pydatetime() - base_file_time).total_seconds() / 3600
@@ -32,6 +35,7 @@ def calculate_time_index(ds, index: str, base_time: str, lead_hours: int) -> int
     time_index = int(np.argmin(np.abs(time_in_hours - valid_hours)))
     logger.info(f"🧭 Closest time index: {time_index}, Available times: {time_in_hours.tolist()}")
     return time_index
+
 
 
 def extract_base_time_from_encoding(ds, index: str) -> datetime:
